requirements.txt
# 智能查询助手依赖包
# ===================

# 核心依赖
gradio>=4.0.0
numpy>=1.20.0
requests>=2.28.1
python-dotenv>=0.21.0

# 嵌入模型支持
sentence-transformers>=2.2.0
torch>=1.9.0
transformers>=4.20.0

# 数据库支持
psycopg2-binary>=2.9.5
pymysql>=1.0.2
sqlite3  # Python内置

# 向量计算
scikit-learn>=1.0.0

# 异步支持
asyncio  # Python内置

# 火山引擎AI支持
# 注意：火山引擎AI API使用标准的HTTP请求，无需额外SDK

# 开发工具（可选）
pytest>=7.0.0
black>=22.0.0
flake8>=5.0.0 